+++ lib/vips/gvalue.rb
@@ -96,6 +96,9 @@ module GObject
             when GINT_TYPE
                 ::GObject::g_value_set_int self, value
 
+            when GUINT64_TYPE
+                ::GObject::g_value_set_uint64 self, value
+
             when GDOUBLE_TYPE
                 ::GObject::g_value_set_double self, value
 
@@ -134,7 +137,7 @@ module GObject
                 value.each {|image| ::GObject::g_object_ref image}
 
             when Vips::BLOB_TYPE
-                len = value.length
+                len = value.bytesize
                 ptr = GLib::g_malloc len
                 Vips::vips_value_set_blob self, GLib::G_FREE, ptr, len
                 ptr.write_bytes value
@@ -174,6 +177,9 @@ module GObject
 
             when GINT_TYPE
                 result = ::GObject::g_value_get_int self
+
+            when GUINT64_TYPE
+                result = ::GObject::g_value_get_uint64 self
 
             when GDOUBLE_TYPE
                 result = ::GObject::g_value_get_double self
@@ -250,6 +256,7 @@ module GObject
 
     attach_function :g_value_set_boolean, [GValue.ptr, :int], :void
     attach_function :g_value_set_int, [GValue.ptr, :int], :void
+    attach_function :g_value_set_uint64, [GValue.ptr, :uint64], :void
     attach_function :g_value_set_double, [GValue.ptr, :double], :void
     attach_function :g_value_set_enum, [GValue.ptr, :int], :void
     attach_function :g_value_set_flags, [GValue.ptr, :uint], :void
@@ -258,6 +265,7 @@ module GObject
 
     attach_function :g_value_get_boolean, [GValue.ptr], :int
     attach_function :g_value_get_int, [GValue.ptr], :int
+    attach_function :g_value_get_uint64, [GValue.ptr], :uint64
     attach_function :g_value_get_double, [GValue.ptr], :double
     attach_function :g_value_get_enum, [GValue.ptr], :int
     attach_function :g_value_get_flags, [GValue.ptr], :int